+++ NEWS.md
@@ -3,7 +3,7 @@
 
 | Date               | News Item                                                                                         | Details |
 | ------------------ | ------------------------------------------------------------------------------------------------- | ------- |
-| 3rd September 2026 | [**recls.Go** 0.1.0-alpha1](https://github.com/synesissoftware/recls.Go) released                 | First functional alpha (Search + Stat) |
+| 3rd September 2026 | [**recls.Go** 0.1.0-alpha1](https://github.com/synesissoftware/recls.Go) released                 | First functional alpha (Search + Stat; `PatternSource`) |
 | 20th August 2026   | [**recls.Go** 0.0.1](https://github.com/synesissoftware/recls.Go/releases/tag/0.0.1) released     | |
 | 20th August 2026   | [**recls.Go** 0.0.0.5](https://github.com/synesissoftware/recls.Go/releases/tag/0.0.0.5) released | |
 | 18th August 2025   | [**recls.Go** 0.0.0.4](https://github.com/synesissoftware/recls.Go/releases/tag/0.0.0.4) released | |